Move static functions away from star class, rename to starboardMessage, requested changes
