+++ src/util/functions.ts
@@ -1,4 +1,4 @@
-import { CaseType } from './constants';
+import { CaseType, IMAGE_LINK_REGEX } from './constants';
 import TimeAgo from 'javascript-time-ago';
 import en from 'javascript-time-ago/locale/en';
 import { guildConfigs } from '../guild/config/guildConfigs';
@@ -8,6 +8,7 @@ import { getPermissionLevel } from './permission/getPermissionLevel';
 import { MinehutClient } from '../client/minehutClient';
 import { cloneDeep } from 'lodash';
 import { PermissionLevel } from './permission/permissionLevel';
+import path from 'path';
 
 TimeAgo.addLocale(en);
 export const ago = new TimeAgo('en-US');
@@ -215,4 +216,44 @@ export function checkString(content: string): CensorCheckResponse | undefined {
 		const match = content.match(regex);
 		if (match) return { rule, match };
 	}
+}
+
+export function emojiEquals(x: any, y: any) {
+	if (typeof x === 'string' && typeof y === 'string') {
+		return x === y;
+	}
+
+	if (typeof x === 'string') {
+		return x === y.name;
+	}
+
+	if (typeof y === 'string') {
+		return x.name === y;
+	}
+
+	return x.identifier === y.identifier;
+}
+
+export function getEmojiFromId(client: MinehutClient, id: string) {
+	if (/^\d+$/.test(id)) return client.emojis.cache.get(id);
+	
+	return id;
+}
+
+export function findImg(msg: Message) {
+	let returnAttachment;
+	const extensions = ['.png', '.jpg', '.jpeg', '.gif', '.webp'];
+
+	const attachment = msg.attachments.find(file =>
+		extensions.includes(path.extname(file.url))
+	);
+	
+	if (attachment) returnAttachment = attachment.url;
+
+	if (!returnAttachment) {
+		const match = msg.content.match(IMAGE_LINK_REGEX);
+		if (match && extensions.includes(path.extname(match[0]))) returnAttachment = match[0];
+	}
+
+	return returnAttachment;
 }
